# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hospital Management API",
        "description": "API for hospital management system",
        "contact": {
            "name": "Swagger API Team"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/doctors": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "add a new doctor to the system",
                "description": "Create Doctor Endpoint.",
                "operationId": "createDoctor",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "password",
                                    "phone_number",
                                    "address",
                                    "gender",
                                    "birth_date",
                                    "departement_id",
                                    "specialization",
                                    "short_description",
                                    "assigned_at",
                                    "rate"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "gender": {
                                        "$ref": "#/components/schemas/Gender"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "departement_id": {
                                        "type": "integer"
                                    },
                                    "specialization": {
                                        "$ref": "#/components/schemas/MedicalSpecialization"
                                    },
                                    "short_description": {
                                        "type": "string"
                                    },
                                    "assigned_at": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "rate": {
                                        "$ref": "#/components/schemas/Rate"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
        "/api/doctors/{id}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "read a doctor info",
                "description": "Read a Specific Doctor Info Endpoint.",
                "operationId": "readDoctor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "doctor's id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Object Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "update doctor info",
                "description": "Update Specific Doctor Info Endpoint.",
                "operationId": "updateDoctor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "doctor's id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "gender": {
                                        "$ref": "#/components/schemas/Gender"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "departement_id": {
                                        "type": "integer"
                                    },
                                    "specialization": {
                                        "$ref": "#/components/schemas/MedicalSpecialization"
                                    },
                                    "short_description": {
                                        "type": "string"
                                    },
                                    "assigned_at": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "rate": {
                                        "$ref": "#/components/schemas/Rate"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Object Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "delete doctor info",
                "description": "Delete Specific Doctor Endpoint.",
                "operationId": "deleteDoctor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "doctor's id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Object Not Found"
                    }
                }
            }
        },
        "/api/doctors/": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "list all doctors",
                "description": "List All Doctors Endpoint.",
                "operationId": "List Doctors",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/doctors/search/": {
            "get": {
                "tags": [
                    "Patient",
                    "Admin"
                ],
                "summary": "Seach on doctors using their full names",
                "description": "Seach on Doctors Endpoint.",
                "operationId": "DoctorSearch",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "doctor's full name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
        "/api/statistics/best_doctors": {
            "get": {
                "tags": [
                    "Patient",
                    "Admin"
                ],
                "summary": "list the best four doctors",
                "description": "Best Doctors Report Endpoint.",
                "operationId": "bestDoctors",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Gender": {
                "title": "Gender Enum",
                "type": "integer",
                "enum": [
                    0,
                    1
                ]
            },
            "MedicalSpecialization": {
                "title": "Medical Specialization Enum",
                "type": "integer",
                "enum": [
                    0,
                    1,
                    2,
                    3,
                    4,
                    5
                ]
            },
            "Rate": {
                "title": "Rate Enum",
                "type": "integer",
                "enum": [
                    0,
                    1,
                    2,
                    3,
                    4
                ]
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}